+++ NSelene/SeleneLocator.cs
@@ -171,6 +171,69 @@ namespace NSelene
         }
     }
 
+    sealed class SCollectionWebElementByItsInnerMatchingConditionSLocator : WebElementSLocator
+    {
+        readonly Condition<SeleneElement> condition;
+        readonly SeleneCollection context;
+        readonly _SeleneSettings_ config;
+        readonly By innerLocator;
+
+        public SCollectionWebElementByItsInnerMatchingConditionSLocator(
+            By innerLocator,
+            Condition<SeleneElement> condition, 
+            SeleneCollection context, 
+            _SeleneSettings_ config
+        )
+        {
+            this.innerLocator = innerLocator;
+            this.condition = condition;
+            this.context = context;
+            this.config = config;
+        }
+
+        public override string Description {
+            get {
+                return $"{this.context}.ElementByIts({this.innerLocator}, {this.condition})";
+            }
+        }
+
+        public override IWebElement Find ()
+        {
+            var webelments = this.context.ActualWebElements;
+
+            Predicate<IWebElement> byCondition = delegate(IWebElement element) {
+                return condition._Predicate(
+                    new SeleneElement(
+                        new WrappedWebElementSLocator(
+                            string.Format("By.Selene: ({0}).FindBy({1})", this.context, condition)
+                            , element), this.config).Element(this.innerLocator)
+                );/* 
+                   * ??? TODO: do we actually need here so meaningful desctiption?
+                   * does it make sense to use it but to put index for each element?
+                   * via using FindIndex ?
+                   */
+            };
+
+            var found = webelments.ToList()
+                                  .Find(byCondition);
+            if (found == null) 
+            {
+                var actualTexts = webelments.ToList().Select(element => element.Text).ToArray();
+                var htmlelements = webelments.ToList().Select(element => element.GetAttribute("outerHTML")).ToArray();
+                throw new NotFoundException("element was not found in collection by condition "
+                                            + condition
+                                            + "\n  Actual visible texts : " + "[" + string.Join(",", actualTexts) + "]"  // TODO: think: this line is actually needed in the case of FindBy(ExactText ...) ... is there any way to add such information not here?
+                                            + "\n  Actual html elements : " + "[" + string.Join(",", htmlelements) + "]" 
+                                            // TODO: should we add here some other info about elements? e.g. visiblitiy?
+                                           );
+                /*
+                     * TODO: think on better messages
+                     */
+            }
+            return found;
+        }
+    }
+
     // TODO: maybe SLocator<ReadOnlyCollection<IWebElement>> ?
     abstract class WebElementsCollectionSLocator : SeleneLocator<ReadOnlyCollection<IWebElement>>
     {